fixed todo

+++ src/DataTypes.java
@@ -5,9 +5,9 @@ public class DataTypes {
     // TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
-        for (Integer x: numbers) {
+        for (long x: numbers) {
             s += x;
         }
         return s;
